ui: do not reset blast method selection if the sequence type info was reset
Sequence type information can be reset when the user re-pastes a new sequence
or edits the pasted sequence (,or types it out by hand). When this happened,
SequenceServer would flush out the blast method selection. Naturally, we
shouldn't make assumptions about the selected blast method unless we know the
sequence type, so we wait until the sequence type information is available
again.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>