ui: handle server errors when blasting over AJAX
Remove spinner and enable BLAST button when AJAX request completes
(successfully or otherwise). And notify the user in the event of a failure.
Earlier the spinner would be stuck up if the server closed connection for some
reason (timeout, blast failed, etc.), leaving the user clueless if it is just a
long blast or some sort of error.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>