reimplement sequence polling functionality as a simple jQuery plugin
Polling the sequence input box for a change was previously tightly coupled with
`SS.blast.signal_sequence_type_changed`, causing the code to look messy and
difficult to understand.  A single method was doing too much -- code smell.
Factoring out the polling logic into a separate utility method simplifies
things a lot.

`$.fn.poll` uses the same logic as before for polling (recursively call itself
after a 100 ms), but the code has been reordered so it is easier to understand.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>