remove .rvmrc
Bundler manages dependencies _really well_.  And that's all we need for
SequenceServer, be it development, or deployment.

I see a `.rvmrc` in the source tree as encouraging the use of RVM.  However, I
find it easier to deal with things without RVM getting in the way.  So I would
rather not-encourage using RVM for SequenceServer.

Signed-off-by: Anurag Priyam <anurag08priyam@gmail.com>